feat(mention): Create generic mention chips for any type of message
- start making use of it for system messages as well
- bump text size from 12sp/bodySmall to 14sp/bodyMedium for system messages (let's not use 12sp, given it is too small and creates A11Y issues for users)

AI-assistant: Claude Code v2.1.126 (claude-sonnet-4-6)
